refactor: cssStyle 분리
