+++ components/src/lib/components/Modal/Modal.tsx
@@ -1,14 +1,25 @@
-import { Children, isValidElement, PropsWithChildren, useEffect } from 'react';
-import { css } from '@emotion/css';
-import { ModalProps, ModalSize, Position } from '../../Modal.type';
+import { PropsWithChildren, useEffect } from 'react';
+import { ModalProps } from '../../Modal.type';
 import ModalBackdrop from '../common/ModalBackdrop';
 import ModalHeader from '../common/ModalHeader';
 import useModalKeyboard from '../../hooks/useModalKeyboard';
 import { ModalProvider, useModalContext } from '../../ModalContext';
 import ConfirmButton from '../common/ConfirmButton';
 import CancelButton from '../common/CancelButton';
+import { ModalFrame } from '../common/cssStyle';
+import { ButtonBar } from '../common/cssStyle';
 
-const Modal = ({ children, position, isOpen, onClose, onAfterOpen, type, size }: ModalProps) => {
+const Modal = ({
+  children,
+  position,
+  isOpen,
+  onClose,
+  onAfterOpen,
+  size,
+  type,
+  showDefaultCancelButton,
+  showDefaultConfirmButton,
+}: ModalProps) => {
   if (!isOpen) return null;
 
   useEffect(() => {
@@ -17,21 +28,56 @@ const Modal = ({ children, position, isOpen, onClose, onAfterOpen, type, size }:
 
   useModalKeyboard(onClose);
 
-  return (
-    <ModalProvider onClose={onClose}>
-      <ModalBackdrop>
-        <div className={ModalFrame(position, size)} data-testid="modal">
-          {children}
-          {Children.toArray(children).every((child) => isValidElement(child) && child.type !== Modal.Actions) && (
-            <div className={ButtonBar}>
-              {(type === 'confirm' || type === 'prompt') && <CancelButton />}
-              <ConfirmButton />
+  switch (type) {
+    case 'confirm':
+      return (
+        <ModalProvider onClose={onClose}>
+          <ModalBackdrop>
+            <div className={ModalFrame(position, size)} data-testid="modal">
+              {children}
+              {(showDefaultCancelButton || showDefaultConfirmButton) && (
+                <div className={ButtonBar}>
+                  {showDefaultCancelButton && <CancelButton />}
+                  {showDefaultConfirmButton && <ConfirmButton />}
+                </div>
+              )}
             </div>
-          )}
-        </div>
-      </ModalBackdrop>
-    </ModalProvider>
-  );
+          </ModalBackdrop>
+        </ModalProvider>
+      );
+    case 'prompt':
+      return (
+        <ModalProvider onClose={onClose}>
+          <ModalBackdrop>
+            <div className={ModalFrame(position, size)} data-testid="modal">
+              {children}
+              {(showDefaultCancelButton || showDefaultConfirmButton) && (
+                <div className={ButtonBar}>
+                  {showDefaultCancelButton && <CancelButton />}
+                  {showDefaultConfirmButton && <ConfirmButton />}
+                </div>
+              )}
+            </div>
+          </ModalBackdrop>
+        </ModalProvider>
+      );
+    default:
+      return (
+        <ModalProvider onClose={onClose}>
+          <ModalBackdrop>
+            <div className={ModalFrame(position, size)} data-testid="modal">
+              {children}
+              {(showDefaultCancelButton || showDefaultConfirmButton) && (
+                <div className={ButtonBar}>
+                  {showDefaultCancelButton && <CancelButton />}
+                  {showDefaultConfirmButton && <ConfirmButton />}
+                </div>
+              )}
+            </div>
+          </ModalBackdrop>
+        </ModalProvider>
+      );
+  }
 };
 
 Modal.Header = function Header({ title, showCloseButton = false }: { title: string; showCloseButton?: boolean }) {
@@ -49,47 +95,3 @@ Modal.Actions = function Actions({ children }: PropsWithChildren) {
 };
 
 export default Modal;
-
-const ModalSizeStyle = {
-  small: css`
-    width: 320px;
-  `,
-  medium: css`
-    width: 480px;
-  `,
-  large: css`
-    width: 600px;
-  `,
-};
-
-const modalPositionStyle = {
-  center: css`
-    border-radius: 8px;
-  `,
-  bottom: css`
-    width: 100dvw;
-    border-radius: 8px 8px 0 0;
-    position: absolute;
-    bottom: 0;
-  `,
-};
-
-const ModalFrame = (position: Position, size: ModalSize) => css`
-  background-color: white;
-  padding: 20px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  gap: 20px;
-  ${ModalSizeStyle[size]}
-  ${modalPositionStyle[position]}
-`;
-
-const ButtonBar = css`
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  gap: 10px;
-  width: 100%;
-`;
